tuned autodownload, save lut at rslcs

+++ bin/framebatch_data_refill.sh
@@ -39,16 +39,31 @@ make_simple_polygon.sh ${frame}-poly.txt
  echo "There are "`cat ${frame}_todown | wc -l`" extra images, not currently existing on CEMS disk"
 # checking if the files from scihub exist in RSLC or SLC folders..
 rm tmp_processed.txt 2>/dev/null
+rm tmp_existing.txt 2>/dev/null
 pom=0
 for rslcdir in RSLC SLC $curdir/$tr/$frame/RSLC; do
  if [ -d $rslcdir ]; then
-  echo "Previous processing exists. Checking"
-  ls $rslcdir >> tmp_processed.txt
+  #echo "Previous processing exists. Checking"
+  ls $rslcdir >> tmp_existing.txt
+  for rslcdatedir in `ls -d $rslcdir/20??????`; do
+   if [ `ls $rslcdatedir/*slc 2>/dev/null | wc -l` -gt 0 ]; then
+    echo $rslcdatedir | rev | cut -d '/' -f1 | rev >> tmp_processed.txt
+   fi
+  done
   pom=1
  fi
 done
-cp tmp_processed.txt tmp_tmp; sort -u tmp_tmp > tmp_processed.txt; rm tmp_tmp
-echo "You have "`cat tmp_processed.txt | wc -l`" already processed images,"
+if [ $pom -eq 1 ]; then
+ cp tmp_processed.txt tmp_tmp; sort -u tmp_tmp > tmp_processed.txt; rm tmp_tmp
+ cp tmp_existing.txt tmp_tmp; sort -u tmp_tmp > tmp_existing.txt; rm tmp_tmp
+ echo "You have "`cat tmp_existing.txt | wc -l`" already processed images,"
+ if [ `cat tmp_existing.txt | wc -l` -gt `cat tmp_processed.txt | wc -l` ]; then
+  echo "the RSLCs are physically existing for "`cat tmp_processed.txt | wc -l`" of them"
+  echo "(those non-existing will be redownloaded and reprocessed)"
+  #do you like the existing X processed paradox?
+  #check this then, you will like it too: https://www.youtube.com/watch?v=2uHNSuGeTpM
+ fi
+fi
  if [ $pom -eq 1 ]; then
   cp ${frame}_db_query.list tmp_dbquery.list
   for file in `cat tmp_dbquery.list`; do 
@@ -60,9 +75,9 @@ echo "You have "`cat tmp_processed.txt | wc -l`" already processed images,"
     sed -i '/'$filedate'/d' ${frame}_todown
    fi
   done
-  rm tmp_processed.txt tmp_dbquery.list
+  rm tmp_processed.txt tmp_dbquery.list tmp_existing.txt 2>/dev/null
  fi
-echo "this means you have now "`cat ${frame}_todown | wc -l`"images to download."
+echo "This means you have now "`cat ${frame}_todown | wc -l`" frame dates to use for download."
  echo "Checking if the NLA-registered files exist on the disk. If not, include them for redownload"
  pom=0
  for file in `cat ${frame}_db_query.list`; do 
@@ -75,10 +90,6 @@ echo "this means you have now "`cat ${frame}_todown | wc -l`"images to download.
  if [ $pom -gt 0 ]; then
   echo "There are "$pom" images that are in NLA but not restored to /neodc"
   echo "Did you run NLA request already?? If not, cancel me and do it first"; sleep 5; 
-  filetodown=`cat ${frame}_todown | wc -l`
-  timetodown=`echo "$filetodown*4500/15/60/60" | bc`
-  echo "ok, nevermind, continuing with "$filetodown" files to download";
-  echo "please note that this can take approx. " $timetodown " hours to finish, hope you run it in tmux or screen????"
  fi
 ## check what really is existing on disk (maybe we missed something?)
  pom=0
@@ -107,20 +118,24 @@ echo "this means you have now "`cat ${frame}_todown | wc -l`"images to download.
      fi
    fi
  done
- if [ $pom -eq 1 ]; then echo "After the checks, there will be "`cat ${frame}_todown | wc -l`" files downloaded"; fi
-
+ filestodown=`cat ${frame}_todown | wc -l`
+ downspeed=5 #MB/s
+ timetodown=`echo "$filestodown*4500/$downspeed/60/60" | bc`
+ echo "After the checks, there will be "$filestodown" files downloaded";
+ echo "please note that this can take approx. " $timetodown " hours to finish, hope you run it in tmux or screen????"
 ## update what is not physically existing on disk 
 ## (we assume users did the nla request before and want to fill all the unavailable data
 ## update 21-02-2019: CEMS disallows secure connection
 ## doing workaround through xfer server (or login server if you do not have approved xfer service)
 xferserver=jasmin-xfer1.ceda.ac.uk
 #testing connection
-ssh -q $xferserver exit
+ssh -q -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no $xferserver exit
 test_conn=$?
 if [ $test_conn -eq 0 ]; then
  echo "will use XFER server to download data"
  sshout=$SLCdir
  sshserver=$xferserver
+ wgetcmd=`which wget_alaska`
 else
  echo "will use login server to download"
  echo "please apply for XFER service in JASMIN website"
@@ -128,19 +143,23 @@ else
  sshdown="ssh cems-login1.cems.rl.ac.uk"
  sshout=~/temp_licsar_down
  mkdir -p $sshout
+ cp `which wget_alaska` $sshout/.
+ wgetcmd=`$sshout/wget_alaska`
  sshserver=cems-login1.cems.rl.ac.uk
  echo "please delete this files as temporary - they were created as workaround to data download. Normally they should be cleaned" > $sshout/README
 fi
-wgetcmd=`which wget_alaska`
 
 #bash workaround to aliases
 shopt -s expand_aliases
-alias sshdown=`echo ssh -q $sshserver "'cd " $sshout "; export LiCSAR_configpath=$LiCSAR_configpath; $wgetcmd '"`
+alias sshdown=`echo ssh -q -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no $sshserver "'cd " $sshout "; export LiCSAR_configpath=$LiCSAR_configpath; $wgetcmd '"`
 
 if [ `cat ${frame}_todown | wc -l` -gt 0 ]; then
  cd $SLCdir
+ count=0
  for x in `cat $BATCH_CACHE_DIR/$frame/${frame}_todown | rev | cut -d '/' -f1 | rev`; do
+  let count=$count+1
   echo "downloading file "$x" from alaska server"
+  echo "( it is file no. "$count" from "$filestodown" )"
   time sshdown $x >/dev/null 2>/dev/null
   if [ ! -f $sshout/$x ]; then
    echo "Some download error appeared, trying again (verbosed)"
